Add files via upload
In this program, we initialize a total and a steps counter. In each step of the while loop, we randomly choose between 2 and 3 and add this to our total. We also increment the steps counter. The loop continues until our total is greater than or equal to 23. After the loop, we print the total and the number of steps it took to get there.